Refactor LiftForm component and add to OfferSupport page
