+++ src/pages/OfferSupportPage.js
@@ -1,0 +1,13 @@
+// import React, { useState } from 'react';
+import LiftForm from '../components/LiftForm';
+
+const OfferSupportPage = () => {
+
+  return (
+    <div>
+      <LiftForm/>
+    </div>
+  );
+}
+
+export default OfferSupportPage
